Move github and sbom functionality out of announce package
