+++ pkg/announce/github/consts.go
@@ -1,0 +1,34 @@
+/*
+Copyright 2024 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package github
+
+// ghPageBody is a generic template to build the GitHub
+// rekease page.
+const ghPageBody = `{{ if .Substitutions.logo }}
+![Logo]({{ .Substitutions.logo }} "Logo")
+{{ end }}
+{{ .Substitutions.intro }}
+{{ if .Substitutions.changelog }}
+See [the CHANGELOG]({{ .Substitutions.changelog }}) for more details.
+{{ end }}
+{{ if .Substitutions.ReleaseNotes }}
+### Release Notes
+
+{{ .Substitutions.ReleaseNotes }}
+{{ end }}
+
+`
