+++ pkg/announce/sbom/options.go
@@ -1,0 +1,44 @@
+/*
+Copyright 2024 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package sbom
+
+const (
+	sbomFileName      = "sbom.spdx"
+	assetDownloadPath = "/releases/download/"
+)
+
+type Asset struct {
+	Path     string // Path where the artifact will be listed
+	ReadFrom string // LocalPath to read the information
+	Label    string // Label for the asset
+}
+
+type SBOMFormat string
+
+const (
+	FormatJSON     SBOMFormat = "json"
+	FormatTagValue SBOMFormat = "tag-value"
+)
+
+type Options struct {
+	ReleaseName   string
+	Repo          string
+	RepoDirectory string
+	Tag           string     // Version Tag
+	Format        SBOMFormat // "tag-value"  | "json"
+	Assets        []Asset
+}
